+++ dvhaedit/dialogs.py
@@ -11,20 +11,14 @@ Classes used to edit pydicom datasets
 #    available at https://github.com/cutright/DVHA-DICOM-Editor
 
 import wx
-from functools import partial
-from os.path import basename, dirname
 import re
-from pubsub import pub
-from queue import Queue
-from threading import Thread
-from time import sleep
-from pydicom.datadict import keyword_dict
 from pydicom.uid import RE_VALID_UID_PREFIX
+from dvhaedit._version import __version__
 from dvhaedit.data_table import DataTable
-from dvhaedit.dicom_editor import TagSearch, DICOMEditor, save_dicom
+from dvhaedit.dicom_editor import TagSearch
 from dvhaedit.dynamic_value import HELP_TEXT
 from dvhaedit.paths import LICENSE_PATH
-from dvhaedit.utilities import save_csv_to_file, get_window_size
+from dvhaedit.utilities import save_csv_to_file
 
 
 class ErrorDialog:
@@ -47,7 +41,7 @@ class AskYesNo(wx.MessageDialog):
     """Simple Yes/No MessageDialog"""
 
     def __init__(self, parent, msg, caption="Are you sure?",
-                 flags=wx.ICON_WARNING | wx.YES | wx.NO | wx.NO_DEFAULT, style=wx.CENTRE):
+                 flags=wx.ICON_WARNING | wx.YES | wx.NO | wx.NO_DEFAULT):
         wx.MessageDialog.__init__(self, parent, msg, caption, flags)
 
 
@@ -97,7 +91,8 @@ class ViewErrorLog(wx.Dialog):
         sizer_wrapper.Add(sizer_buttons, 0, wx.ALIGN_RIGHT | wx.ALL, 5)
 
         self.SetSizer(sizer_wrapper)
-        self.SetSize(get_window_size(0.4, 0.4))
+        self.SetMinSize((700, 600))
+        self.Fit()
         self.Center()
 
     def run(self):
@@ -173,7 +168,8 @@ class TagSearchDialog(wx.Dialog):
         sizer_wrapper.Add(sizer_main, 1, wx.EXPAND | wx.ALL, 5)
 
         self.SetSizer(sizer_wrapper)
-        self.SetSize(get_window_size(0.4, 0.4))
+        self.SetMinSize((700, 400))
+        self.Fit()
         self.Center()
 
     def run(self):
@@ -217,10 +213,10 @@ class TagSearchDialog(wx.Dialog):
 
 class TextViewer(wx.Dialog):
     """Simple dialog to display the LICENSE file and a brief text header in a scrollable window"""
-    def __init__(self, text, title, width=0.3, height=0.6):
+    def __init__(self, text, title, min_size):
         wx.Dialog.__init__(self, None, title=title)
 
-        self.size = get_window_size(width, height)
+        self.SetMinSize(min_size)
 
         self.scrolled_window = wx.ScrolledWindow(self, wx.ID_ANY)
         self.text = wx.StaticText(self.scrolled_window, wx.ID_ANY, text)
@@ -243,7 +239,7 @@ class TextViewer(wx.Dialog):
         sizer_wrapper.Add(self.scrolled_window, 1, wx.EXPAND, 0)
 
         self.SetSizer(sizer_wrapper)
-        self.SetSize(self.size)
+        self.Fit()
         self.Center()
 
     def run(self):
@@ -253,240 +249,18 @@ class TextViewer(wx.Dialog):
 
 class About(TextViewer):
     """Simple dialog to display the LICENSE file and a brief text header in a scrollable window"""
-    def __init__(self, version):
+    def __init__(self):
 
         with open(LICENSE_PATH, 'r', encoding="utf8") as license_file:
             license_text = ''.join([line for line in license_file])
-        license_text = "DVHA DICOM Editor v%s\nedit.dvhanalytics.com\n\n%s" % (version, license_text)
+        license_text = "DVHA DICOM Editor v%s\nedit.dvhanalytics.com\n\n%s" % (__version__, license_text)
 
-        TextViewer.__init__(self, license_text, title='About DVHA DICOM Editor')
+        TextViewer.__init__(self, license_text, title='About DVHA DICOM Editor', min_size=(700, 500))
 
 
 class DynamicValueHelp(TextViewer):
     def __init__(self):
-        TextViewer.__init__(self, HELP_TEXT, title='Dynamic Values', width=0.4)
-
-
-class ProgressFrame(wx.Dialog):
-    """Create a window to display progress and begin provided worker"""
-    def __init__(self, obj_list, action, close_msg, action_msg=None, action_gui_phrase='Processing', title='Progress'):
-        wx.Dialog.__init__(self, None)
-
-        self.close_msg = close_msg
-        self.worker_args = [obj_list, action, action_msg, action_gui_phrase, title]
-        self.action_gui_phrase = action_gui_phrase
-
-        self.gauge = wx.Gauge(self, wx.ID_ANY, 100)
-        self.label = wx.StaticText(self, wx.ID_ANY, "Progress Label:")
-
-        self.__set_properties()
-        self.__do_subscribe()
-        self.__do_layout()
-
-        self.run()
-
-    def run(self):
-        """Initiate layout in GUI and begin thread"""
-        self.Show()
-        ProgressFrameWorker(*self.worker_args)
-
-    def __set_properties(self):
-        width, _ = get_window_size(0.4, 1)
-        self.SetMinSize((width, 100))
-
-    def __do_subscribe(self):
-        pub.subscribe(self.update, "progress_update")
-        pub.subscribe(self.set_title, "progress_set_title")
-        pub.subscribe(self.close, "progress_close")
-
-    @staticmethod
-    def __do_unsubscribe():
-        pub.unsubAll(topicName="progress_update")
-        pub.unsubAll(topicName="progress_set_title")
-        pub.unsubAll(topicName="progress_close")
-
-    def __do_layout(self):
-        sizer_wrapper = wx.BoxSizer(wx.VERTICAL)
-        sizer_objects = wx.BoxSizer(wx.VERTICAL)
-        sizer_objects.Add(self.label, 0, 0, 0)
-        sizer_objects.Add(self.gauge, 0, wx.EXPAND, 0)
-        sizer_wrapper.Add(sizer_objects, 0, wx.ALL | wx.EXPAND, 10)
-        self.SetSizer(sizer_wrapper)
-        self.Fit()
-        self.Layout()
-        self.Center()
-
-    def set_title(self, msg):
-        wx.CallAfter(self.SetTitle, msg)
-
-    def update(self, msg):
-        """
-        Update the progress message and gauge
-        :param msg: a dictionary with keys of 'label' and 'gauge' text and progress fraction, respectively
-        :type msg: dict
-        """
-        label = msg['label']
-        wx.CallAfter(self.label.SetLabelText, label)
-        wx.CallAfter(self.gauge.SetValue, int(100 * msg['gauge']))
-
-    def close(self):
-        """Destroy layout in GUI and send message close message for parent"""
-        wx.CallAfter(pub.sendMessage, self.close_msg)
-        self.__do_unsubscribe()
-        wx.CallAfter(self.Destroy)
-
-
-class ProgressFrameWorker(Thread):
-    def __init__(self, obj_list, action, action_msg, action_gui_phrase, title):
-        Thread.__init__(self)
-
-        pub.sendMessage("progress_set_title", msg=title)
-
-        self.obj_list = obj_list
-        self.obj_count = len(self.obj_list)
-        self.action = action
-        self.action_msg = action_msg
-        self.action_gui_phrase = action_gui_phrase
-
-        self.start()
-
-    def run(self):
-        queue = self.get_queue()
-        worker = Thread(target=self.target, args=[queue])
-        worker.setDaemon(True)
-        worker.start()
-        queue.join()
-        sleep(0.3)  # Allow time for user to see final progress in GUI
-        pub.sendMessage('progress_close')
-
-    def get_queue(self):
-        queue = Queue()
-        for i, obj in enumerate(self.obj_list):
-            msg = {'label': '%s %s of %s' % (self.action_gui_phrase, i + 1, self.obj_count),
-                   'gauge': float(i) / self.obj_count}
-            queue.put((obj, msg))
-        return queue
-
-    def target(self, queue):
-        while queue.qsize():
-            parameters = queue.get()
-            self.do_action(*parameters)
-            queue.task_done()
-
-        msg = {'label': 'Process Complete',
-               'gauge': 1.}
-        pub.sendMessage("progress_update", msg=msg)
-
-    def do_action(self, obj, msg):
-        pub.sendMessage("progress_update", msg=msg)
-
-        result = self.action(obj)
-        if self.action_msg is not None:
-            msg = {'obj': obj, 'data': result}
-            pub.sendMessage(self.action_msg, msg=msg)
-
-
-class ParsingProgressFrame(ProgressFrame):
-    """Create a window to display parsing progress and begin ParseWorker"""
-    def __init__(self, file_paths):
-        ProgressFrame.__init__(self, file_paths, DICOMEditor, close_msg='parse_complete',
-                               action_msg='add_parsed_data', action_gui_phrase='Parsing File',
-                               title='Reading DICOM Headers')
-
-
-class SavingProgressFrame(ProgressFrame):
-    """Create a window to display saving progress and begin SaveWorker"""
-    def __init__(self, data_sets):
-        ProgressFrame.__init__(self, data_sets, save_dicom, close_msg='save_complete',
-                               action_gui_phrase='Saving File',
-                               title='Saving DICOM Data')
-
-
-class RefSyncProgressFrame(ProgressFrame):
-    """Create a window to display Referenced tag syncing progress and begin SaveWorker"""
-    def __init__(self, history, data_sets, check_all_tags):
-        ProgressFrame.__init__(self, history, partial(update_referenced_tags, data_sets, check_all_tags),
-                               close_msg='ref_sync_complete',
-                               action_gui_phrase='Checking References for Tag:',
-                               title='Checking for Referenced Tags')
-
-
-def update_referenced_tags(data_sets, check_all_tags, history_row):
-    keyword, old_value, new_value = tuple(history_row)
-    if "Referenced%s" % keyword in list(keyword_dict):
-        for ds in data_sets:
-            ds.sync_referenced_tag(keyword, old_value, new_value, check_all_tags=check_all_tags)
-
-
-class ValueGenProgressFrame(ProgressFrame):
-    """Create a window to display value generation progress and begin SaveWorker"""
-    def __init__(self, data_sets, value_generator, iteration, total_count):
-
-        ProgressFrame.__init__(self, [data_sets], value_generator,
-                               close_msg='value_gen_complete',
-                               action_msg='add_value_dicts',
-                               action_gui_phrase='File:',
-                               title='Generating Values for Tag %s of %s' % (iteration, total_count))
-
-
-class ApplyEditsProgressFrame(ProgressFrame):
-    """Create a window to display value generation progress and begin SaveWorker"""
-    def __init__(self, data_sets, values_dicts, all_row_data):
-
-        ProgressFrame.__init__(self, [data_sets], partial(apply_edits, values_dicts, all_row_data),
-                               close_msg='do_save_dicom',
-                               action_msg='update_dicom_edits',
-                               action_gui_phrase='',
-                               title='Editing DICOM Data')
-
-
-def apply_edits_callback(iteration, count_total, label):
-    msg = {'label': label,
-           'gauge': iteration / count_total}
-    pub.sendMessage("progress_update", msg=msg)
-
-
-def apply_edits(values_dicts, all_row_data, data_sets):
-    """Apply the tag edits to every file in self.ds, return any errors"""
-    error_log, history = [], []
-    for row in range(len(all_row_data)):
-
-        row_data = all_row_data[row]
-        tag = row_data['tag']
-
-        keyword = row_data['keyword']
-        value_str = row_data['value_str']
-        value_type = row_data['value_type']
-        values_dict = values_dicts[row]
-
-        for i, (file_path, ds) in enumerate(data_sets.items()):
-            label = "Editing %s for file %s of %s" % (keyword, i+1, len(data_sets))
-            apply_edits_callback(i, len(data_sets), label)
-            try:
-                if tag.tag in ds.dcm.keys():
-                    new_value = value_type(values_dict[file_path])
-                    old_value, _ = ds.edit_tag(new_value, tag=tag.tag)
-                    history.append([keyword, old_value, new_value])
-                else:
-                    addresses = ds.find_tag(tag.tag)
-                    if not addresses:
-                        raise Exception
-                    for address in addresses:
-                        new_value = value_type(values_dict[file_path])
-                        old_value, _ = ds.edit_tag(new_value, tag=tag.tag, address=address)
-                        history.append([keyword, old_value, new_value])
-
-            except Exception as e:
-                err_msg = 'KeyError: %s is not accessible' % tag if str(e).upper() == str(tag).upper() else e
-                value = value_str if value_str else '[empty value]'
-                modality = ds.dcm.Modality if hasattr(ds.dcm, 'Modality') else 'Unknown'
-                error_log.append("Directory: %s\nFile: %s\nModality: %s\n\t"
-                                 "Attempt to edit %s to new value: %s\n\t%s\n" %
-                                 (dirname(file_path), basename(file_path), modality, tag, value, err_msg))
-
-    return {'error_log': '\n'.join(error_log),
-            'history': history,
-            'ds': data_sets}
+        TextViewer.__init__(self, HELP_TEXT, title='Dynamic Values', min_size=(672, 420))
 
 
 class AdvancedSettings(wx.Dialog):
@@ -523,7 +297,7 @@ class AdvancedSettings(wx.Dialog):
 
         self.text_ctrl['rand_digits'].SetValue(str(self.options.rand_digits))
 
-        self.SetMinSize(get_window_size(0.4, 0.2))
+        self.SetMinSize((672, 210))
 
     def __do_bind(self):
         self.Bind(wx.EVT_TEXT, self.update_ok_enable, id=self.text_ctrl['rand_digits'].GetId())